Move ManualPatchSet class from gerrit-events to gerrit-trigger module
ManualPatchSet is a special event which is only created by gerrit-trigger
to handle patchset triggered manually from Jenkins UI.
